requirements.txt
SpeechRecognition>=3.10.0
pyttsx3>=2.90
openai>=1.12.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
gTTS>=2.5.0
pygame>=2.6.0
yt-dlp>=2025.0.0
pydub>=0.25.0
